+++ CHANGELOG.md
@@ -6,6 +6,9 @@ und das Projekt nutzt [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 - Konfiguration via config.yaml
+- Unterscheidung Ollama und LMStudio
+- HotReloadTool für LMStudio
+- Modelswitching
 
 ## [0.1.0] – 2025-11-25
 Initialer stabiler Release des Proxys.